+++ bluefin/src/Bluefin/DslBuilderEff.hs
@@ -1,0 +1,11 @@
+-- | Like "Bluefin.DslBuilder", but when you want to be able to run
+-- additional effects as well.
+
+module Bluefin.DslBuilderEff (
+    DslBuilderEff,
+    dslBuilderEff,
+    runDslBuilderEff,
+  )
+where
+
+import Bluefin.Internal.DslBuilderEff